+++ automl_alex/automl_alex.py
@@ -65,6 +65,7 @@ class BestSingleModel(LightGBM):
         auto_parameters=True,
         models_names=None, #list models_names for opt
         feature_selection=True,
+        iteration_check=True,
         verbose=1,
         ):
         '''
@@ -95,6 +96,7 @@ class BestSingleModel(LightGBM):
             early_stoping, 
             feature_selection,
             iterations=iterations,
+            iteration_check=iteration_check,
             verbose=verbose,)
         return(history)
 
@@ -167,6 +169,7 @@ class ModelsReview(BestSingleModel):
         auto_parameters=False,
         feature_selection=True,
         direction=None,
+        iteration_check=False,
         verbose=1,
         models_names=None,
         ):
@@ -212,6 +215,7 @@ class ModelsReview(BestSingleModel):
                         cold_start=50, 
                         opt_lvl=3,
                         feature_selection=feature_selection,
+                        iteration_check=iteration_check,
                         verbose= (lambda x: 0 if x <= 1 else 1)(verbose),
                         )
             if verbose > 0:
@@ -288,29 +292,48 @@ class AutoML(BestSingleModel):
 
         # calc ~ time for opt
         #min_stack_model_timeout = 1000
-        predict_timeout = 30*stack_top # time for predict
+        predict_timeout = int((timeout/100)*20) # time for predict 20% of timeout
         select_models_timeout = (timeout - predict_timeout)
-        model_1_timeout = select_models_timeout - (select_models_timeout//3)
-        model_2_timeout = select_models_timeout//3
-        if (model_1_timeout < 200) or (model_2_timeout < 200):
-            raise Exception(f"Please give me more time to optimize or reduce the number of stack models ('stack_top')")
-
-        # if stack_models_names is None:
-        #     self.stack_models_names = all_models.keys()
-        # else:
-        #     self.stack_models_names = stack_models_names
+        model_2_timeout = select_models_timeout//4
+        model_1_timeout = select_models_timeout - model_2_timeout
+        if (model_1_timeout < 300) or (model_2_timeout < 300):
+            raise Exception("Please, give me more time to optimize!")
 
         ####################################################
         # STEP 1
-        # Model 1
+        # Model 0
+        start_model_0_time = time.time()
         if verbose > 0:
             print('_'*50)
-            print('Step 1: Model 1')
+            print('Step 1: Model 0')
             print('_'*50)
+            time.sleep(0.2) # clean print 
+        # Catboost is good, but it works very slowly, 
+        # so I only make predictions with standard parameters (it works well with standard parameters)
+        model_0 = CatBoost(databunch=self._data, 
+                            cv=5,
+                            score_cv_folds=1,
+                            metric=self.metric,
+                            direction=self.direction,
+                            metric_round=self._metric_round,
+                            combined_score_opt=self._combined_score_opt,
+                            gpu=self._gpu, 
+                            random_state=self._random_state,
+                            type_of_estimator=self.type_of_estimator)
+        
+        self.predicts_model_0_full_x = model_0.predict(on_cv=False,)
+        
+        total_model_0_time = (time.time() - start_model_0_time)
+        model_1_timeout = model_1_timeout - total_model_0_time
+        
+        # Model 1
+        if verbose > 0:
+            print('-'*50)
+            print('Model 1')
             time.sleep(0.2) # clean print 
     
         # Model
-        model = BestSingleModel(databunch=self._data,
+        model_1 = BestSingleModel(databunch=self._data,
                                 opt_lvl=self._opt_lvl,
                                 cv=self._cv,
                                 score_cv_folds = self._score_cv_folds,
@@ -324,7 +347,9 @@ class AutoML(BestSingleModel):
                                 type_of_estimator=self.type_of_estimator)
 
         # Opt
-        history = model.opt(
+        start_opt_model_1_time = time.time()
+        
+        history = model_1.opt(
             timeout=model_1_timeout, 
             early_stoping=early_stoping,
             cold_start=cold_start,
@@ -334,40 +359,84 @@ class AutoML(BestSingleModel):
             score_cv_folds=score_cv_folds,
             auto_parameters=auto_parameters,
             feature_selection=feature_selection,
-            models_names=['LightGBM', 'RandomForest', 'ExtraTrees', 'XGBoost', 'CatBoost'],
+            models_names=['RandomForest', 'ExtraTrees', 'XGBoost', 'LightGBM',],
             verbose= (lambda x: 0 if x <= 1 else 1)(verbose), )
+        
+        total_opt_model_1_time = (time.time() - start_opt_model_1_time)
 
+        # Predict
         history = history.drop_duplicates(subset=['model_score', 'score_std'], keep='last')
-        stack_models_1_cfgs = history.head(stack_top)
 
         if verbose > 0:
             print("\n Predict from Models_1")
             time.sleep(0.2) # clean print 
-        # Predict
-        predicts_1 = model.predict(models_cfgs=stack_models_1_cfgs)
+    
+        # Predict on Full dataset and Time lemit calc
+        start_predict_model_1_time = time.time()
+        self.predicts_model_1_full_x = model_1.predict(models_cfgs=history.head(3), on_cv=False)
+        predict_model_1_time = (time.time() - start_predict_model_1_time)
         
-        self.predicts_model_1_full_x = model.predict(models_cfgs=stack_models_1_cfgs, on_cv=False)
+        if auto_parameters:
+            if verbose > 0:
+                print("\n > Calc predict policy on Models_1:")
+            # Calc predict policy
+            total_time_model_0_1 = (time.time() - start_model_0_time)
+            timeout_predict = (timeout - (total_time_model_0_1 + model_2_timeout + (predict_timeout//3)))
+            iter_predict_time = predict_model_1_time / 3
+            posible_repeats = int(timeout_predict // (iter_predict_time * model_1._cv))
+            n_repeats = 1
+            stack_top_repeats = 1
+
+            if posible_repeats > 1:
+                stack_top_repeats = posible_repeats
+            if posible_repeats > 5:
+                stack_top_repeats = int(posible_repeats//2)
+                n_repeats = 2
+            if posible_repeats > 11:
+                stack_top_repeats = int(posible_repeats//3)
+                n_repeats = 3
+            if posible_repeats > 30:
+                stack_top_repeats = 10
+                n_repeats = 3
+            stack_top = stack_top_repeats
+            
+            if verbose > 0:
+                print(" | posible_repeats: ", posible_repeats, " | stack_top: ", stack_top_repeats, " | n_repeats: ", n_repeats, )
+        
+            # Predict on CV
+            stack_models_1_cfgs = history.head(stack_top_repeats)
+            predicts_1 = model_1.predict(models_cfgs=stack_models_1_cfgs, n_repeats=n_repeats)
+        
+        # Predict on CV
+        else:
+            stack_models_1_cfgs = history.head(stack_top)
+            predicts_1 = model_1.predict(models_cfgs=stack_models_1_cfgs, n_repeats=2)
         
         # Score:
         score_mean_models_1 = self.metric(self._data.y_train, predicts_1['predict_train'].mean())
         if verbose > 0:
             print(f'\n Models_1 Mean {self.metric.__name__} Score Train: ', 
                 round(score_mean_models_1, self._metric_round))
-            print('_'*50)
-            print('Step 2: Model 2')
-            print('_'*50)
-            time.sleep(0.1) # clean print 
         
-
+        end_model_1_time = (time.time() - start_opt_model_1_time)
         #############################################################
         # STEP 2
         # Model 2
+        if verbose > 0:
+            print('-'*50)
+            print('Model 2')
+            time.sleep(0.1) # clean print 
+            
         self.history_trials = []
         self.history_trials_dataframe = pd.DataFrame()
         
+        model_2_timeout = (timeout - end_model_1_time) - predict_timeout
+        if model_2_timeout < 200:
+            model_2_timeout = 200
+        
         model_2 = BestSingleModel(databunch=self._data,
                                   cv=10,
-                                  score_cv_folds = 10,
+                                  score_cv_folds = 4,
                                 metric=self.metric,
                                 direction=self.direction,
                                 metric_round=self._metric_round,
@@ -379,21 +448,25 @@ class AutoML(BestSingleModel):
         # Opt
         history_2 = model_2.opt(
             timeout=model_2_timeout, 
+            early_stoping=early_stoping,
             opt_lvl=3,
-            auto_parameters=False,
+            auto_parameters=True,
             cold_start=25,
             feature_selection=True,
             models_names=['LinearModel', 'MLP',],
+            iteration_check=False,
             verbose= (lambda x: 0 if x <= 1 else 1)(verbose), )
 
         history_2 = history_2.drop_duplicates(subset=['model_score', 'score_std'], keep='last')
         stack_model_2_cfgs = history_2.head(stack_top//2)
+        if stack_top == 1:
+            stack_model_2_cfgs = history_2.head(1)
 
         if verbose > 0:
             print("\n Predict from Models_2")
             time.sleep(0.2) # clean print 
         # Predict
-        predicts_2 = model_2.predict(models_cfgs=stack_model_2_cfgs)
+        predicts_2 = model_2.predict(models_cfgs=stack_model_2_cfgs, n_repeats=1)
         
         # Score:
         score_mean_models_2 = self.metric(self._data.y_train, predicts_2['predict_train'].mean())
@@ -412,7 +485,7 @@ class AutoML(BestSingleModel):
             print(f'\n Mean Models {self.metric.__name__} Score Train: ', \
                 round(score_mean_stack_models, self._metric_round))
             print('_'*50)
-            print('Step 3: Stacking')
+            print('Step 2: Stacking')
             print('_'*50)
             time.sleep(0.1) # clean print 
             
@@ -490,7 +563,9 @@ class AutoML(BestSingleModel):
                 round(final_score_model, self._metric_round))
             time.sleep(0.1) # clean print 
             
-        pred_test = (pred_test * 0.75) + (self.predicts_model_1_full_x['predict_test'].mean() * 0.25)
+        pred_test = (pred_test * 0.7) + \
+            (self.predicts_model_1_full_x['predict_test'].mean() * 0.2) + \
+            (self.predicts_model_0_full_x['predict_test'].mean() * 0.1)
 
         return (pred_test, pred_train)