+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [0.07.26]
+
+### Add
+- Calc predict policy in AutoML
+
+### Fix
+- timelemit in AutoML (deleted Catboost in optimization)
+
+
 ## [0.07.25]
 
 ### Add